Doc

Description

Este módulo contém variadas funções, que foram realizadas através da ficha LI1_3.pdf .

Ficha LI1_3.pdf

Documentation

dobros :: [Float] -> [Float]

Descrição:

A função dobros , recebe uma lista com elementos do tipo Float e retorna uma lista do mesmo tipo, com os seus elementos duplicados.

Código Fonte:

dobros :: [Float] -> [Float]
dobros [] = []
dobros (h:t) = 2*h : dobros t

Exemplos de Utilização:

>>> dobros [1.0,2.0,3.0]
[2.0,4.0,6.0]
>>> dobros [4.0,10.0,8.0]
[8.0,20.0,16.0]

ocorre :: Eq a => a -> [a] -> Int

Descrição:

A função ocorre , recebe um Char e uma lista de elementos do tipo Char, e retorna o número de vezes que esse Char ocorre na String (lista de Char).

Código Fonte:

ocorre :: Eq a => a -> [a] -> Int 
ocorre ch [] = 0 
ocorre ch (h:t) = if ch == h then 1 + ocorre ch t else ocorre ch t 

Exemplos de Utilização:

>>> ocorre 'a' "amanhecer"
2
>>> ocorre '1' "1 + 1 = 1 + 0 + 1"
4

A lista pode escrita desta forma: "amanhecer";

Ou, também pode ser escrita assim: ['a','m','a','n','h','e','c','e','r'];

pmaior :: Int -> [Int] -> Int

Descrição:

A função pmaior , recebe um número n Int e uma lista de elementos do tipo Int, e retorna o primeiro número da lista que é maior que n caso, não encontre retorna n.

Código Fonte:

pmaior :: Int -> [Int] -> Int
pmaior num [] = num 
pmaior num (h:t) = if num < h then h else pmaior num t   

Exemplos de Utilização:

>>> pmaior 4 [2,1,3,4,5,6,10,20]
5
>>> pmaior 6 [2,1,3,4,5,6,10,20]
10

repetidos :: [Int] -> Bool

Descrição:

A função repetidos , recebe uma lista de elementos do tipo Int e verifica se há números repetidos.

Código Fonte:

repetidos :: [Int] -> Bool 
repetidos [] = False
repetidos (h:t) = if ocorre h t > 0 then True else repetidos t  

Exemplos de Utilização:

>>> repetidos [1,2,3,4,4,5]
True
>>> repetidos [1,2,3,4,5,6]
False

tresUlt :: [a] -> [a]

Descrição:

A função tresUlt, recebe uma lista de qualquer tipo, e retorna os 3 últimos elementos dessa mesma lista.

Código Fonte:

tresUlt :: [a] -> [a]
tresUlt [] = []
tresUlt (h:[]) = [h]
tresUlt (x1:x2:[]) = [x1,x2]
tresUlt (x1:x2:x3:[]) = [x1,x2,x3]
tresUlt (h:t) = tresUlt t

Exemplos de Utilização:

>>> tresUlt [1,2,3,4,5,6,7]
[5,6,7]
>>> tresUlt "Bom dia Mundo!"
['d','o','!']

Caso a lista tenha menos de 3 elementos, a função retorna a própria lista.

Funciona para qualquer tipo.

posImpares :: [a] -> [a]

Descrição:

A função posImpares, recebe uma lista de elementos de qualquer tipo, e retorna uma lista com os elementos nas posições ímpares.

Código Fonte:

posImpares :: [a] -> [a]
posImpares (h:t) = aux 1 (h:t)
aux :: Int -> [a] -> [a]
aux n [] = []
aux n (h:t) | mod n 2 == 0 = aux (n+1) t
aux n (h:t) | mod n 2 /= 0 = h:(aux (n+1) t)

Exemplos de Utilização:

>>> posImpares [1,2,3,4,5]
[2,4]
>>> posImpares "Bom dia Mundo!"
"o i ud!"

Nesta função utiliza-se um método diferente, em que criamos outra função, aux para facilitar a função posImpares.

aux :: Int -> [a] -> [a]

somaNeg :: [Int] -> Int

Descrição:

A função somaNeg, recebe uma lista de elementos do tipo Int, e retorna a soma de todos os elementos negativos dessa mesma lista.

Código Fonte:

somaNeg :: [Int] -> Int 
somaNeg [] = 0
somaNeg (h:t) | h>=0 = somaNeg t
somaNeg (h:t) | h<0 = h+somaNeg t

Exemplos de Utilização:

>>> somaNeg [1,2,-3,4,5,-1]
-4
>>> somaNeg [1,2,3,4,5,1]
0

Caso não existam elementos negativos a função somaNeg retorna 0.

soDigitos :: [Char] -> [Char]

minusculas :: [Char] -> Int

type Jogo = (String, Int, String, Int)