Main

Description

Módulo: Main

Título: Problema B

Descrição: Módulo Haskell que dado um tabuleiro já funcional, efetua melhoramentos visuais ao mesmo.

Copyright: Francisco Oliveira a78416@alunos.uminho.pt | | | Vitor Peixoto a79175@alunos.uminho.pt

Resumo: Módulo contendo definições Haskell do Problema B da 1ª Fase do Projeto da disciplina de Laboratórios de Informática I. Este módulo contém funções que melhoram o aspeto visual, nomeadamente, introduzindo as caixas e o Sokoban nas respetivas coordenadas e eliminando os carateres # que são redundantes, ou seja, não influenciem o funcionamento do jogo através do seu mapa.

FUNÇÕES DEFINIDAS NO ESQUELETO INICIAL

Documentation

inStr :: String -> [String]

Função que dada uma String, separa-a onde encontra \n e junta as diferentes String criadas numa lista de String.

>>> "Ola\nMundo"
["Ola","Mundo"]

outStr :: [String] -> String

Função que dada uma lista de String devolve os elementos da lista, agrupados e apenas separados por \n. Função inversa de inStr.

>>> ["Ola","Mundo"]
"Ola\nMundo\n"

main :: IO ()

Função principal (main).

tarefa2 :: [String] -> [String]

Função responsável pela funcionalidade da tarefa, que neste caso é colocar as caixas e o Sokoban nas coordenadas respetivas e eliminar os "#" redundantes.

FUNÇÕES DE MELHORAMENTO VISUAL DO MAPA

funcaoPlantaMelhor :: [String] -> [String]

Dado o tabuleiro com as coordenadas (String) devolve o tabuleiro com o Sokoban nas respetivas coordenadas e representado pelo "o", as caixas também nas respetivas coordenadas representadas por "H" e as caixas quando estão em cima do ponto são representadas por "I".

>>> funcaoPlantaMelhor ["#####","#. .#","#   #","#   #","#####","2 1","1 2","3 2"," "]
["#####","#. .#","#H H#","# o #","#####"]

>>> funcaoPlantaMelhor ["#####","#. .#","#   #","#   #","#####","2 2","1 2","3 3"," "]
["#####","#. I#","#Ho #","#   #","#####"]

funcaoRemoveCard :: [String] -> [String]

Função que remove os cardinais redundantes do tabuleiro, ou seja, aqueles que não são necessários, sem influenciar a jogabilidade do tabuleiro.

>>> funcaoRemoveCard ["########","#. ..###","#### ###","###    #","########"]
["######  ","#. ..#  ","#### ###","  #    #","  ######"]

funcaoRemoveCardCantos :: [String] -> Int -> [String]

Função que seleciona os cardinais que estão nos quatro cantos do tabuleiro e os substitui por pontos de interrogação "?" caso o carater esteja rodeado dentro dos limites do tabuleiro por cardinais "#".

>>> funcaoRemoveCardCantos ["#####","##.##","#   #","## ##","#####"] 4
["?###?","##.##","#   #","## ##","?###?"]

>>> funcaoRemoveCardCantos ["#####","#. .#","#   #","#   #","#####"] 4
["#####","#. .#","#   #","## ##","?###?"]

funcaoRemoveCardLados :: [String] -> [String]

Função que seleciona cada cardinal que está nas bordas do tabuleiro e, caso esteja rodeado de cardinais "#" e/ou "?", substitui-o por um ponto de interrogação "?".

>>> funcaoRemoveCardLados ["?######","###.  #","#    ##","#    ##","######?"]
["??#####","###.  #","#    ##","#    #?","######?"]

>>> funcaoRemoveCardLados ["?####?","##.###","#   .#","##  ##","?####?"]
["?###??","##.###","#   .#","##  ##","?####?"]

funcaoRemoveCardCentro :: [String] -> [String]

Função que seleciona cada cardinal que está nas interior do tabuleiro (ou seja, não está nas bordas do tabuleiro)e, caso esteja rodeado de cardinais "#" e/ou "?" na sua imediação, substitui-o por um ponto de interrogação "?".

>>> funcaoRemoveCardCentro ["######??","#. ..##?","#### ###","?##    #","??######"]
["######??","#. ..#??","#### ###","??#    #","??######"]

funcaoTiraInterrogacao :: [String] -> [String]

Função que substitui os carateres pontos de interrogação "?" colocados anteriormente por espaços " ". Deste modo eliminamos os "#" redundantes e que eram desnecessários ao tabuleiro.

>>> funcaoTiraInterrogacao ["??#####","###.  #","#    ##","#    #?","######?"]
["  #####","###.  #","#    ##","#    # ","###### "]

FUNÇÕES AUXILIARES DE funcaoRemoveCardCentro

funcaoRemoveCardCentroAux :: [String] -> (Int, Int) -> [String]

Função auxiliar de funcaoRemoveCardCentro.

funcaoRemoveCardCentroAux1 :: [String] -> (Int, Int) -> Int -> [String]

Função auxiliar de funcaoRemoveCardCentro.

funcaoRemoveCardCentroAux2 :: [String] -> (Int, Int) -> Int -> [String]

Função auxiliar de funcaoRemoveCardCentro.

funcaoRemoveCardLadosL :: [String] -> Int -> Bool

Função auxiliar de funcaoRemoveCardLados que seleciona cada cardinal que está no lado esquerdo (L - Left) do tabuleiro e, caso esteja rodeado de cardinais "#" e/ou "?", substitui-o por um ponto de interrogação "?".

funcaoRemoveCardLadosR :: [String] -> Int -> Bool

Função auxiliar de funcaoRemoveCardLados que seleciona cada cardinal que está no lado direito (R - Right) do tabuleiro e, caso esteja rodeado de cardinais "#" e/ou "?", substitui-o por um ponto de interrogação "?".

funcaoRemoveCardLadosU :: [String] -> Int -> Bool

Função auxiliar de funcaoRemoveCardLados que seleciona cada cardinal que está na primeira linha (U - Up) do tabuleiro e, caso esteja rodeado de cardinais "#" e/ou "?", substitui-o por um ponto de interrogação "?".

funcaoRemoveCardLadosD :: [String] -> Int -> Bool

Função auxiliar de funcaoRemoveCardLados que seleciona cada cardinal que está na última fila (D - Down) do tabuleiro e, caso esteja rodeado de cardinais "#" e/ou "?", substitui-o por um ponto de interrogação "?".

updateTAB :: [String] -> (Int, Int) -> [String]

Função auxiliar de funcaoRemoveCardLados e funcaoRemoveCardCantos que dado um tabuleiro (em lista de String) e a coordenada de uma caixa (par de Int) devolve o tabuleiro com a caixa inserida na coordenada referida.

>>> updateTAB ["#####","#. .#","#   #","#   #","#####"] (2,2)
["#####","#. .#","# # #","#   #","#####"]

>>> updateTAB ["#####","#. .#","#   #","#   #","#####"] (4,1)
["#####","#. .#","#   #","#   #","#####"]

SEPARAR TABULEIRO DAS COORDENADAS

sepTABCoord :: [String] -> [String]

Função auxiliar de FuncaoPlantaMelhor que separa o tabuleiro das coordenadas. Dá apenas o tabuleiro.

>>> sepTABCoord ["#####","#. .#","#   #","#   #","#####","2 1","1 2","3 2"," "]
["#####","#. .#","#   #","#   #","#####"]

sepTabCOORD :: [String] -> [String]

Separação das coordenadas do tabuleiro. Dá apenas as coordenadas da localização do Sokoban e das caixas.

>>> sepTabCOORD ["#####","#. .#","#   #","#   #","#####","2 1","1 2","3 2"," "]
["2 1","1 2","3 2"]

sepCOORD1 :: [String] -> [String]

Separa as coordenadas uma a uma.

>>> sepCOORD1 ["2 1","1 2","3 2"]
["2","1","1","2","3","2"]

sepCOORD2 :: [String] -> [(Int, Int)]

Após separar as coordenadas uma a uma, juntamo-las em pares de inteiros.

>>> sepCOORD2 ["2","1","1","2","3","2"]
[(2,1),(1,2),(3,2)]

sepTabCOORDpares :: [String] -> [(Int, Int)]

Função auxiliar de funcaoPlantaMelhor que junta as três anteriores funções e executa todas apenas nesta.

>>> sepTabCOORDpares ["#####","#. .#","#   #","#   #","#####","2 1","1 2","3 2"," "]
[(2,1),(1,2),(3,2)]

FUNÇÕES AUXILIARES

updateTABcaixas :: [String] -> (Int, Int) -> [String]

Função auxiliar de FuncaoPlantaMelhor que dado um tabuleiro e a coordenada de uma caixa, coloca essa caixa no tabuleiro, que fica representada pelo carater "H" caso a caixa se localize num espaço vazio ou pelo carater "I" caso fique em cima de qualquer outro carater, ou seja, o ponto.

>>> updateTABcaixas ["#####","#. .#","#   #","#   #","#####"] (1,2)
["#####","#. .#","#H  #","#   #","#####"]

>>> updateTABcaixas ["#####","#. .#","#   #","#   #","#####"] (1,1)
["#####","#I .#","#   #","#   #","#####"]

updateTABsokoban :: [String] -> (Int, Int) -> [String]

Função auxiliar de FuncaoPlantaMelhor que dado um tabuleiro e a coordenada do Sokoban, coloca-o no tabuleiro, na sua respetiva coordenada. O Sokoban está representado pelo carater "o".

>>> updateTABsokoban ["#####","#. .#","#   #","#   #","#####"] (1,1)
["#####","#o .#","#   #","#   #","#####"]

>>> updateTABsokoban ["#####","#. .#","#   #","#   #","#####"] (3,2)
["#####","#. .#","#  o#","#   #","#####"]

charNaPosicao :: [String] -> (Int, Int) -> Char

Função auxiliar das funções auxiliares de funcaoRemoveCardLados que dado um tabuleiro e as coordenadas de uma caixa ou Sokoban diz o que está nessa coordenada. Se estiver vazio, está disponível para receber uma caixa ou o Sokoban.

>>> charNaPosicao ["#####","#. .#","#   #","#   #","#####"] (2,2)
' '

>>> charNaPosicao ["#####","#. .#","#   #","#   #","#####"] (4,1)
'#'

updateTABAux :: String -> Int -> String

Função auxiliar de updateTAB.

>>> updateTABAux "#   #" 1
"##  #"

updateTABcaixasAux :: String -> Int -> String

Função auxiliar de updateTABcaixas que define a letra a colocar na coordenada da caixa. Se no espaço estiver o carater " ", coloca "H". Se não for vazio, coloca o carater "I".

updateTABsokobanAux :: String -> Int -> String

Função auxiliar de updateTABsokoban que lê a coordenada do Sokoban e o coloca no tabuleiro representado pelo carater "o".

funcaoTiraInterrogacaoAux :: String -> String

Função auxiliar de funcaoTiraInterrogacao.

FUNÇÕES PREDEFINIDAS

nub1 :: Eq a => [a] -> [a]

Função predefinida que elimina as repetições dos elementos de uma lista.

>>> nub1 [1,3,1,5,3,4,2,3,1,3]
[1,3,5,4,2]