Main

Description

Módulo: Main

Título: Problema F

Descrição: Módulo Haskell que, com auxílio da biblioteca Gloss cria o jogo Sokoban mas desta vez em modo gráfico 2D.

Copyright: Francisco Oliveira a78416@alunos.uminho.pt | | | Vitor Peixoto a79175@alunos.uminho.pt

Resumo: Módulo contendo definições Haskell do Problema F da 2ª Fase do Projeto da disciplina de Laboratórios de Informática I. Este módulo contém funções que desenha o tabuleiro numa janela gráfica com o auxílio de Pictures importadas dos bitmaps. O mapa é atualizado a cada comando que lhe é dado pela função joga. Os comandos são dados na função reageEvento que executa alterações na janela de acordo com comandos dados no teclado ou rato do computador.

FUNÇÕES DEFINIDAS NO ESQUELETO INICIAL

Documentation

inStr :: String -> [String]

Função que dada uma String, separa-a onde encontra \n e junta as diferentes String criadas numa lista de String.

>>> "Ola\nMundo"
["Ola","Mundo"]

outStr :: [String] -> String

Função que dada uma lista de String devolve os elementos da lista, agrupados e apenas separados por \n. Função inversa de inStr.

>>> ["Ola","Mundo"]
"Ola\nMundo\n"

main :: IO ()

Função principal (main).

type Estado = ([String], [Picture], String, Int, String)

Tipo que representa o estado do jogo.

joga :: Estado -> (Estado -> Picture) -> (Event -> Estado -> Estado) -> IO ()

Função responsável pela atualização contínua do jogo a cada comando que é introduzido na função reageEvento.

MAPAS DO JOGO

tabuleiros :: [String]

Função tabuleiros que aglomera os diversos níveis do jogo numa lista de String.

mapa0 :: String

mapa1 :: String

mapa2 :: String

mapa3 :: String

mapa4 :: String

mapa5 :: String

mapa6 :: String

mapa7 :: String

mapa8 :: String

mapa9 :: String

FUNÇÕES DE DESENHO DOS MAPAS

desenhaJogo :: Estado -> Picture

Funcão que desenha todas as Pictures que devem aparecer na janela. No caso de Pictures estáticas a sua inserção é basica. No entanto, no caso das Pictures que integram o mapa de jogo, é necessário recorrer a uma função auxiliar complexa, a função desenhaJogoAux.

desenhaJogoAux :: ([[Char]], [Picture]) -> (Float, Float) -> [Picture]

Função desenhaJogoAux que é a responsável por inserir na janela as Pictures que constituem o mapa de jogo e que estão em constante mudança pela introdução de moves.

reageEvento :: Event -> Estado -> Estado

Funcão que lê os comandos introduzidos pelo utilizador quer através do teclado quer através de cliques com o rato. Estes comandos podem ser desde a seleção de níveis e retroceder um movimento, até próprio movimento do Sokoban, fazendo desta função extremamente importante.

tarefa4 :: [String] -> [String]

Função que dado um tabuleiro retribui esse mesmo tabuleiro com o Sokoban e as caixas colocadas nas respetivas coordenadas e após os movimentos introduzidos.

>>> tarefa4 ["#####","#. .#","#   #","#   #","#####","2 1","1 2","3 2","RUDLLU",""]
["#####","#I I#","#o  #","#   #","#####"]

>>> tarefa4 ["#####","#. .#","#   #","#   #","#####","2 1","1 2","3 2","RUDLL",""]
["#####","#. I#","#H  #","#o  #","#####"]

FUNÇÕES DE SEQUÊNCIA DE COMANDOS

rinceAndRepeat :: [String] -> String -> [String]

Função que inserido um tabuleiro e uma sucessão de comandos executa os comandos até acabar a sequência de comandos ou até o jogo estar ganho, ou seja, até não haver nenhum H no tabuleiro. É devolvido o tabuleiro com o Sokoban e as caixas nas coordenadas após os movimentos, quer o jogo esteja ganho ou não.

>>> rinceAndRepeat ["#####","#    #","#  .#","#####","1 1","2 1",""] "R"
["#####","#   #","# oI#","#####"]

>>> rinceAndRepeat ["#####","#    #","#  .#","#####","1 1","2 1",""] "UDUR"
["#####","# o #","# H.#","#####"]

FUNÇÕES IMPORTADAS DE OUTRAS TAREFAS

detetaPontos :: [String] -> [(Int, Int)]

Função que procura pontos e guarda as suas coordenadas.

>>> detetaPontos ["#####","# H #","#oI.#","#####"]
[(2,1),(3,1)]

>>> detetaPontos ["#####","#   #","#oH.#","#####"]
[(3,1)]

updateTABpontos :: [String] -> [(Int, Int)] -> [String]

Função que recebe um tabuleiro e as coordenadas dos pontos e os coloca nas respetivas coordenadas. Se o espaço estiver ocupado por um carater qualquer, não coloca o ponto, no entanto ele está lá, debaixo desse carater. Se estiver vazio, coloca um ponto.

>>> updateTABpontos ["#####","#   #","# o #","#####"] [(1,1)]
["#####","#   #","#.o #","#####"]

>>> updateTABpontos ["#####","#   #","#o  #","#####"] [(1,1)]
["#####","#   #","#o  #","#####"]

updateTABponto :: [String] -> (Int, Int) -> [String]

Função auxiliar de updateTABpontos.

updateTABpontoAux :: String -> Int -> String

Função auxiliar de updateTABponto.

vitoria :: [String] -> String

Função que testa se o conjunto de comandos introduzidos levou à vitória, ou seja, se não há nenhum carater H no tabuleiro. Caso haja, ainda não está terminado o jogo, ainda há caixas por meter nos pontos.

>>> vitoria ["#####","#   #","#oH.#","#####"]
"INCOMPLETO"

>>> vitoria ["#####","#   #","# oI#","#####"]
"FIM"

tarefa3' :: [String] -> (Int, Int) -> String -> (([String], (Int, Int)), Int)

Função importada da Tarefa C que dado um tabuleiro com as coordenadas do Sokoban, caixas e um comando de movimentos (U,D,R,L) devolve a nova coordenada do Sokoban resultante desse movimento.

movORnot :: [String] -> (Int, Int) -> (Int, Int) -> (([String], (Int, Int)), Int)

Função que movimenta o Sokoban através da introdução de um tabuleiro com as coordenadas do Sokoban e das coordenadas definidas para movimentar o Sokoban ((0,1) para cima, (0,-1) para baixo, (1,0) para a direita, (-1,0) para a esquerda) e devolve o novo tabuleiro com as alterações efetuadas (ou não) e a coordenada do Sokoban. Devolve ainda 1 ou 0 que significa se o comando inserido efetuou alguma alteração no tabuleiro. Este dado será usado para o tick_count.

>>> movORnot ["#####","#   #","#oH.#","#####"] (1,1) (-1,0)
((["#####","#   #","#oH.#","#####"],(1,1)),0)

>>> >>> movORnot ["#####","#   #","#oH.#","#####"] (1,1) (0,1)
((["#####","#o  #","# H.#","#####"],(1,2)),1) 

movTOempty :: [String] -> (Int, Int) -> (Int, Int) -> [String]

Função que dado o tabuleiro, a coordenada do Sokoban e o comando do movimento, devolve o tabuleiro com o Sokoban na sua nova posição. Esta função é auxiliar de movORnot pelo que só vai ser usada para mover o Sokoban quando o movimento o leva para uma coordenada vazia ou um ponto.

>>> movTOempty ["#####","#   #","#oH.#","#####"] (1,1) (0,1)
["#####","#o  #","# H.#","#####"]

movTObox :: [String] -> (Int, Int) -> (Int, Int) -> [String]

Função que dado o tabuleiro, a coordenada do Sokoban e o comando do movimento, devolve o tabuleiro com o Sokoban e a caixa que empurrou na sua nova posição. Esta função é auxiliar de movORnot pelo que só vai ser usada para mover o Sokoban quando o movimento o leva a empurrar uma caixa.

>>> movTObox ["#####","#   #","#oH.#","#####"] (1,1) (1,0)
["#####","#   #","# oI#","#####"]

charNaPosicao :: [String] -> (Int, Int) -> Char

Função auxiliar de movORnot que dado um tabuleiro e as coordenadas de uma caixa ou Sokoban diz o que está nessa coordenada. Se estiver vazio, está disponível para receber uma caixa ou o Sokoban.

>>> charNaPosicao ["#####","#. .#","#   #","#   #","#####"] (2,2)
' '

>>> charNaPosicao ["#####","#. .#","#   #","#   #","#####"] (4,1)
'#'

FUNÇÃO PLANTAMELHOR E SUAS AUXILIARES

funcaoPlantaMelhor :: [String] -> [String]

Dado o tabuleiro com as coordenadas (String) devolve o tabuleiro com o Sokoban nas respetivas coordenadas e representado pelo "o", as caixas também nas respetivas coordenadas representadas por "H" e as caixas quando estão em cima do ponto são representadas por "I".

>>> funcaoPlantaMelhor ["#####","#. .#","#   #","#   #","#####","2 1","1 2","3 2"," "]
["#####","#. .#","#H H#","# o #","#####"]

>>> funcaoPlantaMelhor ["#####","#. .#","#   #","#   #","#####","2 2","1 2","3 3"," "]
["#####","#. I#","#Ho #","#   #","#####"]

FUNÇÕES UPDATE

updateTABcaixas :: [String] -> (Int, Int) -> [String]

Função auxiliar de funcaoPlantaMelhor que dado um tabuleiro e a coordenada de uma caixa, coloca essa caixa no tabuleiro, que fica representada pelo carater "H" caso a caixa se localize num espaço vazio ou pelo carater "I" caso fique em cima de qualquer outro carater, ou seja, o ponto.

>>> updateTABcaixas ["#####","#. .#","#   #","#   #","#####"] (1,2)
["#####","#. .#","#H  #","#   #","#####"]

>>> updateTABcaixas ["#####","#. .#","#   #","#   #","#####"] (1,1)
["#####","#I .#","#   #","#   #","#####"]

updateTABsokoban :: [String] -> (Int, Int) -> [String]

Função auxiliar de funcaoPlantaMelhor que dado um tabuleiro e a coordenada do Sokoban, coloca-o no tabuleiro, na sua respetiva coordenada. O Sokoban está representado pelo carater "o".

>>> updateTABsokoban ["#####","#. .#","#   #","#   #","#####"] (1,1)
["#####","#o .#","#   #","#   #","#####"]

>>> updateTABsokoban ["#####","#. .#","#   #","#   #","#####"] (3,2)
["#####","#. .#","#  o#","#   #","#####"]

updateTABempty :: [String] -> (Int, Int) -> [String]

Função que após a movimentação do Sokoban, coloca um espaço vazio na sua posição original.

updateTABcaixasAux :: String -> Int -> String

Função auxiliar de updateTABcaixas que define a letra a colocar na coordenada da caixa. Se no espaço estiver o carater " ", coloca "H". Se não for vazio, coloca o carater "I".

updateTABsokobanAux :: String -> Int -> String

Função auxiliar de updateTABsokoban que lê a coordenada do Sokoban e o coloca no tabuleiro representado pelo carater "o".

updateTAB :: [String] -> (Int, Int) -> [String]

Função auxiliar de funcaoRemoveCardLados e funcaoRemoveCardCantos que dado um tabuleiro (em lista de String) e a coordenada de uma caixa (par de Int) devolve o tabuleiro com a caixa inserida na coordenada referida.

>>> updateTAB ["#####","#. .#","#   #","#   #","#####"] (2,2)
["#####","#. .#","# # #","#   #","#####"]

>>> updateTAB ["#####","#. .#","#   #","#   #","#####"] (4,1)
["#####","#. .#","#   #","#   #","#####"]

updateTABAux :: String -> Int -> String

Função auxiliar de updateTAB.

>>> updateTABAux "#   #" 1
"##  #"

updateTABemptyAux :: String -> Int -> String

Função auxiliar de updateTABempty.

algarismos :: Char -> Bool

Função auxiliar de testeCOORDso2numeros que dada uma lista de Char diz se essa lista tem apenas números. Qualquer outro tipo de carater devolve Falso.

>>> algarismos [1,3,5,a,6,b]
False

>>> algarismos [1,3,7,4,2]
True

SEPARAR TABULEIRO DAS COORDENADAS

sepTABCoord :: [String] -> [String]

Função auxiliar de funcaoPlantaMelhor que separa o tabuleiro das coordenadas. Dá apenas o tabuleiro.

>>> sepTABCoord ["#####","#. .#","#   #","#   #","#####","2 1","1 2","3 2"," "]
["#####","#. .#","#   #","#   #","#####"]

sepTabCOORD :: [String] -> [String]

Separação das coordenadas do tabuleiro. Dá apenas as coordenadas da localização do Sokoban e das caixas.

>>> sepTabCOORD ["#####","#. .#","#   #","#   #","#####","2 1","1 2","3 2"," "]
["2 1","1 2","3 2"]

sepCOORD1 :: [String] -> [String]

Separa as coordenadas uma a uma.

>>> sepCOORD1 ["2 1","1 2","3 2"]
["2","1","1","2","3","2"]

sepCOORD2 :: [String] -> [(Int, Int)]

Após separar as coordenadas uma a uma, juntamo-las em pares de inteiros.

>>> sepCOORD2 ["2","1","1","2","3","2"]
[(2,1),(1,2),(3,2)]

sepTabCOORDpares :: [String] -> [(Int, Int)]

Função auxiliar de funcaoPlantaMelhor que junta as três anteriores funções e executa todas apenas nesta.

>>> sepTabCOORDpares ["#####","#. .#","#   #","#   #","#####","2 1","1 2","3 2"," "]
[(2,1),(1,2),(3,2)]

FUNÇÕES DE REMOÇÃO DE CARDINAIS REDUNDANTES

funcaoRemoveCard :: [String] -> [String]

Função que remove os cardinais redundantes do tabuleiro, ou seja, aqueles que não são necessários, sem influenciar a jogabilidade do tabuleiro.

>>> funcaoRemoveCard ["########","#. ..###","#### ###","###    #","########"]
["######  ","#. ..#  ","#### ###","  #    #","  ######"]

funcaoRemoveCardCantos :: [String] -> Int -> [String]

Função que seleciona os cardinais que estão nos quatro cantos do tabuleiro e os substitui por pontos de interrogação "?" caso o carater esteja rodeado dentro dos limites do tabuleiro por cardinais "#".

>>> funcaoRemoveCardCantos ["#####","##.##","#   #","## ##","#####"] 4
["?###?","##.##","#   #","## ##","?###?"]

>>> funcaoRemoveCardCantos ["#####","#. .#","#   #","#   #","#####"] 4
["#####","#. .#","#   #","## ##","?###?"]

funcaoRemoveCardLados :: [String] -> [String]

Função que seleciona cada cardinal que está nas bordas do tabuleiro e, caso esteja rodeado de cardinais "#" e/ou "?", substitui-o por um ponto de interrogação "?".

>>> funcaoRemoveCardLados ["?######","###.  #","#    ##","#    ##","######?"]
["??#####","###.  #","#    ##","#    #?","######?"]

>>> funcaoRemoveCardLados ["?####?","##.###","#   .#","##  ##","?####?"]
["?###??","##.###","#   .#","##  ##","?####?"]

funcaoRemoveCardCentro :: [String] -> [String]

Função que seleciona cada cardinal que está nas interior do tabuleiro (ou seja, não está nas bordas do tabuleiro)e, caso esteja rodeado de cardinais "#" e/ou "?" na sua imediação, substitui-o por um ponto de interrogação "?".

>>> funcaoRemoveCardCentro ["######??","#. ..##?","#### ###","?##    #","??######"]
["######??","#. ..#??","#### ###","??#    #","??######"]

funcaoTiraInterrogacao :: [String] -> [String]

Função que substitui os carateres pontos de interrogação "?" colocados anteriormente por espaços " ". Deste modo eliminamos os "#" redundantes e que eram desnecessários ao tabuleiro.

>>> funcaoTiraInterrogacao ["??#####","###.  #","#    ##","#    #?","######?"]
["  #####","###.  #","#    ##","#    # ","###### "]

FUNÇÕES AUXILIARES DE funcaoRemoveCardCentro

funcaoRemoveCardCentroAux :: [String] -> (Int, Int) -> [String]

Função auxiliar de funcaoRemoveCardCentro.

funcaoRemoveCardCentroAux1 :: [String] -> (Int, Int) -> Int -> [String]

Função auxiliar de funcaoRemoveCardCentro.

funcaoRemoveCardCentroAux2 :: [String] -> (Int, Int) -> Int -> [String]

Função auxiliar de funcaoRemoveCardCentro.

funcaoRemoveCardLadosL :: [String] -> Int -> Bool

Função auxiliar de funcaoRemoveCardLados que seleciona cada cardinal que está no lado esquerdo (L - Left) do tabuleiro e, caso esteja rodeado de cardinais "#" e/ou "?", substitui-o por um ponto de interrogação "?".

funcaoRemoveCardLadosR :: [String] -> Int -> Bool

Função auxiliar de funcaoRemoveCardLados que seleciona cada cardinal que está no lado direito (R - Right) do tabuleiro e, caso esteja rodeado de cardinais "#" e/ou "?", substitui-o por um ponto de interrogação "?".

funcaoRemoveCardLadosU :: [String] -> Int -> Bool

Função auxiliar de funcaoRemoveCardLados que seleciona cada cardinal que está na primeira linha (U - Up) do tabuleiro e, caso esteja rodeado de cardinais "#" e/ou "?", substitui-o por um ponto de interrogação "?".

funcaoRemoveCardLadosD :: [String] -> Int -> Bool

Função auxiliar de funcaoRemoveCardLados que seleciona cada cardinal que está na última fila (D - Down) do tabuleiro e, caso esteja rodeado de cardinais "#" e/ou "?", substitui-o por um ponto de interrogação "?".

funcaoTiraInterrogacaoAux :: String -> String

Função auxiliar de funcaoTiraInterrogacao.